:sparkles: feat(ChatBTPScripts): Add transformation script for yaml files
A new script, TakeAFileAsInputTransformAndCreateAFileAsOutput.sh, has been added to the ChatBTPScripts directory. This script is designed to transform yaml files, specifically Helm files, by removing certain variables and replacing them with a template. The script also removes any unused variables from the yaml file. This change was made to streamline the process of transforming yaml files and to ensure that the output is clean and free of unnecessary information.